requirements.txt
click
prometheus-client
pyyaml
